requirements.txt
requests>=2.31.0
rich>=13.7.0
python-whois>=0.8.0
dnspython>=2.4.0
python-dotenv>=1.0.0
